test(tour): add unit test on utils dom
