+++ src/components/tour/utils/__mocks__/is-visible.ts
@@ -1,0 +1,5 @@
+import { vi } from 'vitest'
+
+export const isVisible = vi.fn(async () => true)
+
+export const focus = vi.fn()
